add: split check_and_inference var_decl function
