+++ src/type_parser/check_and_inference/mod.rs
@@ -1,1 +1,3 @@
 pub mod check_and_inference;
+mod variable_declaration;
+mod _struct;
