+++ src/type_parser/check_and_inference/variable_declaration/mod.rs
@@ -1,0 +1,1 @@
+pub mod variable_declaration;
